# YJIT: Allow parallel scanning for JIT-compiled code