--- a/gc/default/default.c
+++ b/gc/default/default.c
@@ -7068,6 +7068,8 @@ gc_update_references(rb_objspace_t *objspace)
 {
     objspace->flags.during_reference_updating = true;
 
+    rb_gc_before_updating_jit_code();
+
     struct heap_page *page = NULL;
 
     for (int i = 0; i < HEAP_COUNT; i++) {
@@ -7101,6 +7103,8 @@ gc_update_references(rb_objspace_t *objspace)
             table
         );
     }
+
+    rb_gc_after_updating_jit_code();
 
     objspace->flags.during_reference_updating = false;
 }