@wks 2025-07-01
There is an alternative solution for `Block::gc_obj_offsets`.  Currently, YJIT embeds object references inside the function.  The x86_64 backend uses 64-bit immediates, and the ARM64 backend emits the code sequence `LDR`, `B` and `.data`, where the `B` instruction jumps over the data.  In both cases, we need to record the offsets or the addresses of those embedded references.

An alternative solution is to put all the object reference constants *after* the function, like this:

```asm
jit_compiled_func1:
  ...
  ldr x1, ref1 ; Load an object reference constant
  ldr x2, ref2 ; Load another object reference constant
  ...
  ldr x3, ref3 ; Load another object reference constant
  ...
  ret
  .align 3  
ref1:
  .xword 0x123456789abcdef0  ; The first embedded reference
ref2:
  .xword 0x123456789abcdf40  ; The second embedded reference
ref3:
  .xword 0x123456789abcdf80  ; The third embedded reference
```

If we use this approach for both x86_64 and ARM64, then we can eliminate `Block::gc_obj_offsets`, and we only need to record the address of label `ref1` and the number of object references used by that function.  That will further simplify the object reference scanning operation because all references are stored contiguously.  But that requires more changes to the YJIT codegen.