@wks 2025-07-02
Here are the results of running yjit-bench.  I ran on my laptop with ArchLinux.  Both the master branch and this PR are built with `configure --prefix=$PWD/install --disable-install-doc && make install -j`.

Some benchmarks are faster, and some are slower.  `fannkuchredux` has a 10.8% slowdown, and it is reproducible.  I don't know what was the cause.

```
master: ruby 3.5.0dev (2025-07-01T11:28:47Z master 9d080765cc) +PRISM [x86_64-linux]
pr: ruby 3.5.0dev (2025-07-01T12:00:40Z feature/ruby-yjit-.. f54b3c5a79) +PRISM [x86_64-linux]

-----------------  -----------  ----------  -------  ----------  ----------  ---------
bench              master (ms)  stddev (%)  pr (ms)  stddev (%)  pr 1st itr  master/pr
activerecord       184.5        0.4         179.2    0.8         1.037       1.029    
chunky-png         397.4        0.1         398.1    0.1         1.006       0.998    
erubi-rails        740.7        0.2         738.6    0.5         0.913       1.003    
hexapdf            1348.0       1.5         1342.7   0.8         0.998       1.004    
liquid-c           41.2         7.8         37.9     5.2         1.055       1.087    
liquid-compile     33.4         5.5         31.4     3.9         0.982       1.062    
liquid-render      89.5         1.8         85.7     1.5         1.026       1.044    
lobsters           630.5        0.5         627.7    0.6         1.000       1.004    
mail               82.1         6.2         80.9     5.2         1.015       1.015    
psych-load         1260.4       0.2         1289.3   0.3         0.978       0.978    
railsbench         1609.1       0.6         1598.1   0.6         1.025       1.007    
rubocop            94.2         0.7         88.8     0.6         1.020       1.061    
ruby-lsp           96.5         0.2         94.4     0.3         1.052       1.023    
sequel             36.8         0.4         36.8     1.3         1.015       1.000    
binarytrees        176.2        0.3         171.7    2.4         0.996       1.027    
blurhash           165.9        0.2         160.2    0.2         1.034       1.036    
erubi              94.2         0.5         93.1     1.4         1.075       1.011    
etanni             134.5        0.4         135.0    0.6         1.025       0.996    
fannkuchredux      248.5        0.1         278.6    1.2         0.890       0.892    
fluentd            217.7        0.4         209.5    0.2         1.020       1.039    
graphql            166.7        0.1         166.5    0.1         1.009       1.001    
graphql-native     297.4        0.1         295.5    0.2         1.024       1.007    
lee                544.6        0.4         544.3    0.6         1.000       1.000    
matmul             271.5        0.2         269.1    0.4         1.008       1.009    
nbody              60.6         0.3         60.3     0.3         1.003       1.005    
nqueens            113.9        0.2         112.3    0.3         1.014       1.015    
optcarrot          2988.9       0.7         3002.4   0.7         0.997       0.996    
protoboeuf         93.0         0.6         91.5     0.3         1.036       1.016    
protoboeuf-encode  75.3         0.3         74.4     0.3         1.013       1.012    
rack               23.6         0.8         23.6     0.7         1.008       0.998    
ruby-json          166.4        0.1         165.5    0.8         1.013       1.006    
rubyboy            2737.4       0.2         2771.8   0.2         0.991       0.988    
rubykon            558.7        1.5         555.9    1.4         0.986       1.005    
sudoku             280.9        0.3         291.8    0.3         0.964       0.963    
tinygql            364.3        0.2         368.0    0.6         0.989       0.990    
30k_ifelse         394.0        0.2         384.3    0.7         1.059       1.025    
30k_methods        320.1        1.6         310.5    0.2         1.022       1.031    
cfunc_itself       44.4         1.9         43.1     2.4         1.061       1.031    
fib                100.8        0.4         100.6    0.4         1.002       1.001    
getivar            45.6         0.6         45.9     0.7         0.992       0.994    
keyword_args       124.0        0.3         123.7    1.0         1.004       1.002    
loops-times        505.9        0.1         496.0    0.3         1.019       1.020    
object-new         40.9         0.2         40.6     0.3         1.015       1.007    
respond_to         102.2        0.6         101.8    1.2         1.007       1.005    
ruby-xor           54.5         0.7         54.1     0.7         0.995       1.007    
setivar            42.8         1.3         42.6     0.6         1.004       1.004    
setivar_object     44.7         0.6         44.6     0.5         0.991       1.001    
setivar_young      44.7         0.6         44.8     0.6         0.999       0.998    
str_concat         31.6         2.6         31.4     2.2         1.002       1.004    
throw              11.3         2.2         11.2     0.7         1.005       1.008    
-----------------  -----------  ----------  -------  ----------  ----------  ---------
```