@wks 2025-07-15
Since an alternative solution https://github.com/ruby/ruby/pull/13843 is merged, we no longer need this PR.